Trainer: time-free monotonic scoring, and auto-cycle-with-Pause on misses
- Scoring is move-quality only, no time: +1 for a good move, -0.5 for an
  inaccuracy, -1 for a blunder (grades are only {-2,-1,+1,+2} in practice, so
  this is monotonic in grade). grading.score_attempt drops the time penalty;
  the per-move score and running total become -1..+1. Think time is still
  measured and shown ("took Xs"), just not scored. TRN-SCORE / NFR-CLOCK reworded.
- Auto mode keeps cycling on a miss too — right advances after ~0.5s, wrong
  after ~2s — but a Pause button stops the timer so you can study a mistake as
  long as you like (resume via Next).

Drops grading.time_penalty and its tests; attempts.final_score holds the score.

+++ src/chesstrain/db.py
@@ -548,7 +548,7 @@ def get_grade(conn: sqlite3.Connection, epd: str) -> sqlite3.Row | None:
 # --- attempts (trainer history) -------------------------------------------
 def insert_attempt(conn: sqlite3.Connection, *, epd: str, source: str,
                    played_uci: str, grade: int, time_taken_s: float,
-                   time_penalty: int, final_score: int, tc_class: str,
+                   time_penalty: int, final_score: float, tc_class: str,
                    ts: float) -> None:
     conn.execute(
         "INSERT INTO attempts(epd, source, played_uci, grade, time_taken_s, "
